change HomeGallery component name
